謎のJavaScriptバグ解消Ver

+++ Assets/GravityPainter/System/UI/MainScene/Script/Gpt_RestEnemyUI.cs
@@ -155,7 +155,7 @@ public class Gpt_RestEnemyUI : MonoBehaviour
         //Debug.Log("fff]"+tenPoint);
         //Debug.Log("ggg]" + onePoint);
 
-        switch (playState.PlayerColor)
+        switch ((int)playState.PlayerColor)
         {
             case 0:
                 countTen.sprite = Count[0];
